[feature] 픽 목록 편집 모드

+++ app/src/main/java/com/squirtles/musicroad/pick/DetailPickScreen.kt
@@ -41,6 +41,8 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.luminance
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.lerp
@@ -53,12 +55,14 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import com.squirtles.domain.model.Pick
 import com.squirtles.musicroad.R
+import com.squirtles.musicroad.common.DialogTextButton
+import com.squirtles.musicroad.common.HorizontalSpacer
+import com.squirtles.musicroad.common.MessageAlertDialog
 import com.squirtles.musicroad.common.VerticalSpacer
 import com.squirtles.musicroad.media.PlayerServiceViewModel
 import com.squirtles.musicroad.pick.PickViewModel.Companion.DEFAULT_PICK
 import com.squirtles.musicroad.pick.components.CircleAlbumCover
 import com.squirtles.musicroad.pick.components.CommentText
-import com.squirtles.musicroad.pick.components.DeletePickDialog
 import com.squirtles.musicroad.pick.components.DetailPickTopAppBar
 import com.squirtles.musicroad.pick.components.MusicVideoKnob
 import com.squirtles.musicroad.pick.components.PickInformation
@@ -66,6 +70,7 @@ import com.squirtles.musicroad.pick.components.SongInfo
 import com.squirtles.musicroad.pick.components.music.MusicPlayer
 import com.squirtles.musicroad.pick.components.music.visualizer.BaseVisualizer
 import com.squirtles.musicroad.ui.theme.Black
+import com.squirtles.musicroad.ui.theme.Primary
 import com.squirtles.musicroad.ui.theme.White
 import com.squirtles.musicroad.videoplayer.MusicVideoScreen
 import kotlinx.coroutines.launch
@@ -272,14 +277,32 @@ fun DetailPickScreen(
     }
 
     if (showDeletePickDialog) {
-        DeletePickDialog(
+        MessageAlertDialog(
             onDismissRequest = {
                 showDeletePickDialog = false
             },
-            onDeletion = {
-                showDeletePickDialog = false
-                pickViewModel.deletePick(pickId)
-            }
+            title = stringResource(R.string.delete_pick_dialog_title),
+            body = stringResource(R.string.delete_pick_dialog_body),
+            buttons = {
+                DialogTextButton(
+                    onClick = {
+                        showDeletePickDialog = false
+                    },
+                    text = stringResource(R.string.delete_pick_dialog_cancel)
+                )
+
+                HorizontalSpacer(8)
+
+                DialogTextButton(
+                    onClick = {
+                        showDeletePickDialog = false
+                        pickViewModel.deletePick(pickId)
+                    },
+                    text = stringResource(R.string.delete_pick_dialog_delete),
+                    textColor = Primary,
+                    fontWeight = FontWeight.Bold
+                )
+            },
         )
     }
